renamed .vue, .cy.ts
